server.tsのリファクタ
